+++ op-chain-ops/genesis/config.go
@@ -39,6 +39,8 @@ type DeployConfig struct {
 	L2OutputOracleOwner              common.Address `json:"l2OutputOracleOwner"`
 	L2OutputOracleGenesisL2Output    common.Hash    `json:"l2OutputOracleGenesisL2Output"`
 
+	SystemConfigOwner common.Address `json:"systemConfigOwner"`
+
 	L1BlockTime                 uint64         `json:"l1BlockTime"`
 	L1GenesisBlockTimestamp     hexutil.Uint64 `json:"l1GenesisBlockTimestamp"`
 	L1GenesisBlockNonce         hexutil.Uint64 `json:"l1GenesisBlockNonce"`
@@ -67,10 +69,10 @@ type DeployConfig struct {
 	L2CrossDomainMessengerOwner common.Address `json:"l2CrossDomainMessengerOwner"`
 	OptimismBaseFeeRecipient    common.Address `json:"optimismBaseFeeRecipient"`
 	OptimismL1FeeRecipient      common.Address `json:"optimismL1FeeRecipient"`
-	GasPriceOracleOwner         common.Address `json:"gasPriceOracleOwner"`
-	GasPriceOracleOverhead      uint           `json:"gasPriceOracleOverhead"`
-	GasPriceOracleScalar        uint           `json:"gasPriceOracleScalar"`
-	GasPriceOracleDecimals      uint           `json:"gasPriceOracleDecimals"`
+
+	GasPriceOracleOwner    common.Address `json:"gasPriceOracleOwner"`
+	GasPriceOracleOverhead uint64         `json:"gasPriceOracleOverhead"`
+	GasPriceOracleScalar   uint64         `json:"gasPriceOracleScalar"`
 
 	DeploymentWaitConfirmations int `json:"deploymentWaitConfirmations"`
 
@@ -172,10 +174,8 @@ func NewL2StorageConfig(config *DeployConfig, block *types.Block, l2Addrs *L2Add
 		"msgNonce":         0,
 	}
 	storage["GasPriceOracle"] = state.StorageValues{
-		"_owner":   config.GasPriceOracleOwner,
-		"overhead": config.GasPriceOracleOverhead,
-		"scalar":   config.GasPriceOracleScalar,
-		"decimals": config.GasPriceOracleDecimals,
+		// TODO: remove this in the future
+		"_owner": config.GasPriceOracleOwner,
 	}
 	storage["L1Block"] = state.StorageValues{
 		"number":         block.Number(),
@@ -183,6 +183,9 @@ func NewL2StorageConfig(config *DeployConfig, block *types.Block, l2Addrs *L2Add
 		"basefee":        block.BaseFee(),
 		"hash":           block.Hash(),
 		"sequenceNumber": 0,
+		"batcherHash":    config.BatchSenderAddress.Hash(),
+		"l1FeeOverhead":  config.GasPriceOracleOverhead,
+		"l1FeeScalar":    config.GasPriceOracleScalar,
 	}
 	storage["LegacyERC20ETH"] = state.StorageValues{
 		"bridge":      predeploys.L2StandardBridge,