+++ op-node/rollup/derive/channel_bank.go
@@ -4,10 +4,11 @@ import (
 	"context"
 	"io"
 
-	"github.com/ethereum-optimism/optimism/op-node/eth"
-	"github.com/ethereum-optimism/optimism/op-node/rollup"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/log"
+
+	"github.com/ethereum-optimism/optimism/op-node/eth"
+	"github.com/ethereum-optimism/optimism/op-node/rollup"
 )
 
 type NextFrameProvider interface {
@@ -158,7 +159,7 @@ func (cb *ChannelBank) NextData(ctx context.Context) ([]byte, error) {
 	}
 }
 
-func (cb *ChannelBank) Reset(ctx context.Context, base eth.L1BlockRef) error {
+func (cb *ChannelBank) Reset(ctx context.Context, base eth.L1BlockRef, _ eth.SystemConfig) error {
 	cb.channels = make(map[ChannelID]*Channel)
 	cb.channelQueue = make([]ChannelID, 0, 10)
 	return io.EOF