+++ op-node/rollup/derive/l1_block_info.go
@@ -6,15 +6,21 @@ import (
 	"fmt"
 	"math/big"
 
-	"github.com/ethereum-optimism/optimism/op-bindings/predeploys"
-	"github.com/ethereum-optimism/optimism/op-node/eth"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
+
+	"github.com/ethereum-optimism/optimism/op-bindings/predeploys"
+	"github.com/ethereum-optimism/optimism/op-node/eth"
+)
+
+const (
+	L1InfoFuncSignature = "setL1BlockValues(uint64,uint64,uint256,bytes32,uint64,bytes32,uint256,uint256)"
+	L1InfoArguments     = 8
+	L1InfoLen           = 4 + 32*L1InfoArguments
 )
 
 var (
-	L1InfoFuncSignature    = "setL1BlockValues(uint64,uint64,uint256,bytes32,uint64)"
 	L1InfoFuncBytes4       = crypto.Keccak256([]byte(L1InfoFuncSignature))[:4]
 	L1InfoDepositerAddress = common.HexToAddress("0xdeaddeaddeaddeaddeaddeaddeaddeaddead0001")
 	L1BlockAddress         = predeploys.L1BlockAddr
@@ -29,10 +35,14 @@ type L1BlockInfo struct {
 	// Not strictly a piece of L1 information. Represents the number of L2 blocks since the start of the epoch,
 	// i.e. when the actual L1 info was first introduced.
 	SequenceNumber uint64
+	// BatcherHash version 0 is just the address with 0 padding to the left.
+	BatcherAddr   common.Address
+	L1FeeOverhead eth.Bytes32
+	L1FeeScalar   eth.Bytes32
 }
 
 func (info *L1BlockInfo) MarshalBinary() ([]byte, error) {
-	data := make([]byte, 4+32+32+32+32+32)
+	data := make([]byte, L1InfoLen)
 	offset := 0
 	copy(data[offset:4], L1InfoFuncBytes4)
 	offset += 4
@@ -49,11 +59,17 @@ func (info *L1BlockInfo) MarshalBinary() ([]byte, error) {
 	copy(data[offset:offset+32], info.BlockHash.Bytes())
 	offset += 32
 	binary.BigEndian.PutUint64(data[offset+24:offset+32], info.SequenceNumber)
+	offset += 32
+	copy(data[offset+12:offset+32], info.BatcherAddr[:])
+	offset += 32
+	copy(data[offset:offset+32], info.L1FeeOverhead[:])
+	offset += 32
+	copy(data[offset:offset+32], info.L1FeeScalar[:])
 	return data, nil
 }
 
 func (info *L1BlockInfo) UnmarshalBinary(data []byte) error {
-	if len(data) != 4+32+32+32+32+32 {
+	if len(data) != L1InfoLen {
 		return fmt.Errorf("data is unexpected length: %d", len(data))
 	}
 	var padding [24]byte
@@ -76,6 +92,12 @@ func (info *L1BlockInfo) UnmarshalBinary(data []byte) error {
 	if !bytes.Equal(data[offset:offset+24], padding[:]) {
 		return fmt.Errorf("l1 info sequence number exceeds uint64 bounds: %x", data[offset:offset+32])
 	}
+	offset += 32
+	info.BatcherAddr.SetBytes(data[offset+12 : offset+32])
+	offset += 32
+	copy(info.L1FeeOverhead[:], data[offset:offset+32])
+	offset += 32
+	copy(info.L1FeeScalar[:], data[offset:offset+32])
 	return nil
 }
 
@@ -88,13 +110,16 @@ func L1InfoDepositTxData(data []byte) (L1BlockInfo, error) {
 
 // L1InfoDeposit creates a L1 Info deposit transaction based on the L1 block,
 // and the L2 block-height difference with the start of the epoch.
-func L1InfoDeposit(seqNumber uint64, block eth.BlockInfo) (*types.DepositTx, error) {
+func L1InfoDeposit(seqNumber uint64, block eth.BlockInfo, sysCfg eth.SystemConfig) (*types.DepositTx, error) {
 	infoDat := L1BlockInfo{
 		Number:         block.NumberU64(),
 		Time:           block.Time(),
 		BaseFee:        block.BaseFee(),
 		BlockHash:      block.Hash(),
 		SequenceNumber: seqNumber,
+		BatcherAddr:    sysCfg.BatcherAddr,
+		L1FeeOverhead:  sysCfg.Overhead,
+		L1FeeScalar:    sysCfg.Scalar,
 	}
 	data, err := infoDat.MarshalBinary()
 	if err != nil {
@@ -120,8 +145,8 @@ func L1InfoDeposit(seqNumber uint64, block eth.BlockInfo) (*types.DepositTx, err
 }
 
 // L1InfoDepositBytes returns a serialized L1-info attributes transaction.
-func L1InfoDepositBytes(seqNumber uint64, l1Info eth.BlockInfo) ([]byte, error) {
-	dep, err := L1InfoDeposit(seqNumber, l1Info)
+func L1InfoDepositBytes(seqNumber uint64, l1Info eth.BlockInfo, sysCfg eth.SystemConfig) ([]byte, error) {
+	dep, err := L1InfoDeposit(seqNumber, l1Info, sysCfg)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create L1 info tx: %w", err)
 	}