+++ packages/contracts-bedrock/hardhat.config.ts
@@ -203,6 +203,10 @@ const config: HardhatUserConfig = {
       default: 2,
     },
 
+    systemConfigOwner: {
+      type: 'address',
+    },
+
     // Optional L1 genesis block values. These must ONLY be used by the L1 genesis config script.
     // Not all deployments may create a new L1 chain, but instead attach to an existing L1 chain, like Goerli.
     // -------------------------------------------------
@@ -352,7 +356,7 @@ const config: HardhatUserConfig = {
   solidity: {
     compilers: [
       {
-        version: '0.8.10',
+        version: '0.8.15',
         settings: {
           optimizer: { enabled: true, runs: 10_000 },
         },